dataflow/requirements.txt
apache-beam[gcp]==2.46.0
google-cloud-storage==2.8.0
pymongo==3.9.0
pymongo[srv]
pandas==1.5.3
setuptools~=67.7.2
mysql-connector-python==8.0.23
pycryptodome==3.17
google-cloud-secret-manager==2.16.1